Location Aria

+++ app/src/main/java/com/example/smartalert/Alert.java
@@ -56,7 +56,7 @@ public class Alert {
         return longitude;
     }
 
-    public void setLongitude(int longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
 
@@ -64,7 +64,7 @@ public class Alert {
         return latitude;
     }
 
-    public void setLatitude(int latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
